requirements.txt
streamlit>=1.35.0
pandas>=2.2.2
openai>=1.30.1
pandasai>=1.6.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
plotly>=5.22.0
Pillow>=10.3.0
matplotlib>=3.7.2
